refactor(profile): Improve robustness of Learning Profile CRUD ops

+++ backend/hilbert-main/src/main/java/com/hilbert/features/learningprofile/model/LearningLanguageData.java
@@ -1,6 +1,5 @@
 package com.hilbert.features.learningprofile.model;
 
-import com.hilbert.shared.common.enums.DifficultyLevel;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -15,9 +14,9 @@ public class LearningLanguageData {
 
     private LocalDateTime startDate;
     private LocalDateTime lastActiveDate;
-    private Integer totalLearningDays;
+    private Long totalLearningDays;
     private Double totalXp;
 
     private List<Long> readArticleIds;
-    private Integer vocabularySize;
+    private Long vocabularySize;
 }
